ref: Rename app component
