+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './components/app/app.module';
+import App from './components/app/app';
 
 import './index.css';
